oppdaterte oppg.5

+++ oppgave_5.py
@@ -6,6 +6,6 @@ person3 = 2.5
 person4 = 21
 person5 = 0
 #regner ut gjennomsnittet
-gjennomsnittet = ((person1+person2+person3+person4+person5)/personer)
+gjennomsnittet = int((person1+person2+person3+person4+person5)/personer)
 #printer gjennomsnittet
 print(gjennomsnittet)
